requirements.txt
alpaca-trade-api==3.0.0
pandas==2.0.0
numpy==1.24.0
plotly==5.14.0
python-dotenv==1.0.0
